requirements.txt
Pillow==8.3.2
lzma==0.5.0